refactor: extract auth form shared logic
